:pick: Separate demo `main` from src file
This is so that downstream users can simply include `biski64.c` (unity
build-style) to get the functions they want without fighting the `main`
that it used to come with.
